Use new reactivity system in sycamore
