+++ packages/sycamore-reactive/Cargo.toml
@@ -16,6 +16,7 @@ version = "0.5.2"
 [dependencies]
 indexmap = {version = "1.7.0", features = ["std"]}
 serde = {version = "1.0", optional = true}
+slotmap = "1.0"
 smallvec = "1.6"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
