--- packages/sycamore-reactive2/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-authors = ["Luke Chu <37006668+lukechu10@users.noreply.github.com>"]
-categories = ["gui", "wasm", "web-programming"]
-description = "A VDOM-less web library with fine grained reactivity"
-edition = "2018"
-homepage = "https://github.com/sycamore-rs/sycamore"
-keywords = ["wasm", "gui", "reactive"]
-license = "MIT"
-name = "sycamore-reactive2"
-readme = "../../README.md"
-repository = "https://github.com/sycamore-rs/sycamore"
-version = "0.5.2"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-indexmap = {version = "1.7.0", features = ["std"]}
-serde = {version = "1.0", optional = true}
-slotmap = "1.0"
-smallvec = "1.6"
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-web-sys = {version = "0.3", features = ["console"]}
