--- packages/sycamore-reactive2/src/lib.rs
@@ -1,15 +1,0 @@
-//! Reactive primitives for Sycamore.
-//!
-//! For more information, read the section on [Reactivity](https://sycamore-rs.netlify.app/docs/basics/reactivity)
-//! in the Sycamore Book.
-
-#![deny(missing_docs)]
-
-mod context;
-pub mod effect;
-mod iter;
-pub mod scope;
-pub mod signal;
-
-pub use context::*;
-pub use iter::*;
