+++ packages/sycamore/tests/ssr/main.rs
@@ -2,55 +2,61 @@ use sycamore::prelude::*;
 
 #[test]
 fn hello_world() {
-    let node = template! {
-        p { "Hello World!" }
-    };
+    let _ = create_root(|| {
+        let node = template! {
+            p { "Hello World!" }
+        };
 
-    assert_eq!(sycamore::render_to_string(|| node), "<p>Hello World!</p>");
+        assert_eq!(sycamore::render_to_string(|| node), "<p>Hello World!</p>");
+    });
 }
 
 #[test]
 fn reactive_text() {
-    let count = Signal::new(0);
+    let _ = create_root(|| {
+        let (count, set_count) = create_signal(0);
 
-    let node = cloned!((count) => template! {
-        p { (count.get()) }
+        let node = template! {
+            p { (count.get()) }
+        };
+
+        let node1 = node.clone();
+        assert_eq!(sycamore::render_to_string(|| node1), "<p>0</p>");
+
+        set_count.set(1);
+        assert_eq!(sycamore::render_to_string(|| node), "<p>1</p>");
     });
-
-    assert_eq!(
-        sycamore::render_to_string(cloned!((node) => move || node)),
-        "<p>0</p>"
-    );
-
-    count.set(1);
-    assert_eq!(sycamore::render_to_string(|| node), "<p>1</p>");
 }
 
 #[test]
 fn self_closing_tag() {
-    let node = template! {
-        div {
-            input
-            input(value="a")
-        }
-    };
+    let _ = create_root(|| {
+        let node = template! {
+            div {
+                input
+                input(value="a")
+            }
+        };
 
-    assert_eq!(
-        sycamore::render_to_string(|| node),
-        "<div><input /><input value=\"a\" /></div>"
-    )
+        assert_eq!(
+            sycamore::render_to_string(|| node),
+            "<div><input /><input value=\"a\" /></div>"
+        );
+    });
 }
 
 #[test]
 fn fragments() {
-    let node = template! {
-        p { "1" }
-        p { "2" }
-        p { "3" }
-    };
+    let _ = create_root(|| {
+        let node = template! {
+            p { "1" }
+            p { "2" }
+            p { "3" }
+        };
 
-    assert_eq!(
-        sycamore::render_to_string(|| node),
-        "<p>1</p><p>2</p><p>3</p>"
-    );
+        assert_eq!(
+            sycamore::render_to_string(|| node),
+            "<p>1</p><p>2</p><p>3</p>"
+        );
+    });
 }
